requirements.txt
flask>=2.0.1
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.2
tensorflow-macos>=2.9.0
tensorflow-metal>=0.5.0
coremltools>=6.3.0
plotly>=5.3.1
python-dateutil>=2.8.2
tqdm>=4.65.0
joblib>=1.2.0
SQLAlchemy>=1.4.36
flask_sqlalchemy>=3.0.2
gunicorn>=20.1.0
python-dotenv>=0.19.0